requirements.txt
python-telegram-bot==8.1.1
xkcd==2.4.2
